Change visibility values to actually observed visibilities (#106)
The previous value set was the flag value set of clang/gcc, but
that does not match what we observe in practice from outputs like
`nm -gC` or manpages like e.g. on ELF.

The new values are local (for hidden symbols), global (for symbols
visible across files), and weak (for weak symbols, possibly not
available on all platforms).